Fix paths for analytics client tests
