+++ __tests__/lib/analytics/clients/google-analytics.js
@@ -7,14 +7,22 @@ import url from 'url';
 /**
  * Internal dependencies
  */
-import Tracks from 'lib/analytics/tracks';
+import GoogleAnalytics from 'lib/analytics/clients/google-analytics';
 
-describe( 'lib/analytics/tracks', () => {
+describe( 'lib/analytics/google-analytics', () => {
 	const {
 		protocol: endpointProtocol,
 		host: endpointHost,
 		path: endpointPath
-	} = url.parse( Tracks.ENDPOINT );
+	} = url.parse( GoogleAnalytics.ENDPOINT );
+
+	const env = {
+		app: {
+			name: 'vip',
+			version: '1.0',
+		},
+		userAgent: 'vip-cli'
+	};
 
 	const buildNock = () => {
 		return nock( `${ endpointProtocol }//${ endpointHost }` )
@@ -25,14 +33,17 @@ describe( 'lib/analytics/tracks', () => {
 
 	describe( '.send()', () => {
 		it( 'should correctly construct remote request', () => {
-			const tracksClient = new Tracks( 123, 'vip', '', {
-				userAgent: 'vip-cli'
-			} );
+			const gaClient = new GoogleAnalytics( 'UA-123', 456, env );
 
 			const params = { extra: 'param' };
 
-			const expectedBody = 'commonProps%5B_ui%5D=123' +
-				'&commonProps%5B_ut%5D=vip' +
+			const expectedBody = 'v=1' +
+				'&tid=UA-123' +
+				'&cid=456' +
+				'&aip=1' +
+				'&ds=cli' +
+				'&an=vip' +
+				'&av=1.0' +
 				'&extra=param';
 
 			buildNock()
@@ -44,43 +55,53 @@ describe( 'lib/analytics/tracks', () => {
 					expect( requestBody ).toEqual( expectedBody );
 				} );
 
-			return tracksClient.send( params );
+			return gaClient.send( params );
 		} );
 	} );
 
 	describe( '.trackEvent()', () => {
-		it( 'should pass event details to request', () => {
-			const tracksClient = new Tracks( 123, 'vip', 'prefix_', {} );
+		it( 'should pass minimum event details to request', () => {
+			const gaClient = new GoogleAnalytics( 'UA-123', 456, env );
 
 			const eventName = 'clickButton';
 			const eventDetails = {
-				buttonName: 'deploy',
+				category: 'cat',
 			};
 
-			const expectedBodyMatch = 'events%5B0%5D%5B_en%5D=prefix_clickButton' +
-				'&events%5B0%5D%5BbuttonName%5D=deploy';
+			const expectedBodyMatch = '&t=event' +
+				'&ea=clickButton' +
+				'&ec=cat';
 
 			buildNock()
 				.reply( ( uri, requestBody ) => {
 					expect( requestBody ).toContain( expectedBodyMatch );
 				} );
 
-			return tracksClient.trackEvent( eventName, eventDetails );
+			return gaClient.trackEvent( eventName, eventDetails );
 		} );
 
-		it( 'should ignore prefix if already set for event name', () => {
-			const tracksClient = new Tracks( 123, 'vip', 'existingprefix_', {} );
+		it( 'should pass all event details to request', () => {
+			const gaClient = new GoogleAnalytics( 'UA-123', 456, env );
 
-			const eventName = 'existingprefix_clickButton';
+			const eventName = 'clickButton';
+			const eventDetails = {
+				category: 'cat',
+				label: 'lab',
+				value: 1,
+			};
 
-			const expectedBodyMatch = 'events%5B0%5D%5B_en%5D=existingprefix_clickButton';
+			const expectedBodyMatch = '&t=event' +
+				'&ea=clickButton' +
+				'&ec=cat' +
+				'&el=lab' +
+				'&ev=1';
 
 			buildNock()
 				.reply( ( uri, requestBody ) => {
 					expect( requestBody ).toContain( expectedBodyMatch );
 				} );
 
-			return tracksClient.trackEvent( eventName, {} );
+			return gaClient.trackEvent( eventName, eventDetails );
 		} );
 	} );
 } );
